move checks to engine
