+++ api/src/opentrons/protocol_api/core/engine/protocol.py
@@ -4,7 +4,6 @@ from typing import Dict, Optional, Type, Union, List, Tuple, TYPE_CHECKING
 
 from opentrons.protocol_engine.commands import LoadModuleResult
 from opentrons_shared_data.deck.dev_types import DeckDefinitionV5, SlotDefV3
-from opentrons.protocol_engine.resources import deck_configuration_provider
 from opentrons_shared_data.labware.labware_definition import LabwareDefinition
 from opentrons_shared_data.labware.dev_types import LabwareDefinition as LabwareDefDict
 from opentrons_shared_data.pipette.dev_types import PipetteNameType
@@ -410,7 +409,6 @@ class ProtocolCore(
 
         robot_type = self._engine_client.state.config.robot_type
         normalized_deck_slot = deck_slot.to_equivalent_for_robot_type(robot_type)
-        self._ensure_module_location(normalized_deck_slot, module_type)
 
         result = self._engine_client.load_module(
             model=EngineModuleModel(model),
@@ -622,30 +620,6 @@ class ProtocolCore(
         return (
             self._engine_client.state.addressable_areas.get_staging_slot_definitions()
         )
-
-    def _ensure_module_location(
-        self, slot: DeckSlotName, module_type: ModuleType
-    ) -> None:
-        if self._engine_client.state.config.robot_type == "OT-2 Standard":
-            slot_def = self.get_slot_definition(slot)
-            compatible_modules = slot_def["compatibleModuleTypes"]
-            if module_type.value not in compatible_modules:
-                raise ValueError(
-                    f"A {module_type.value} cannot be loaded into slot {slot}"
-                )
-        else:
-            cutout_fixture_id = ModuleType.to_module_fixture_id(module_type)
-            module_fixture = deck_configuration_provider.get_cutout_fixture(
-                cutout_fixture_id,
-                self._engine_client.state.addressable_areas.state.deck_definition,
-            )
-            cutout_id = self._engine_client.state.addressable_areas.get_cutout_id_by_deck_slot_name(
-                slot
-            )
-            if cutout_id not in module_fixture["mayMountTo"]:
-                raise ValueError(
-                    f"A {module_type.value} cannot be loaded into slot {slot}"
-                )
 
     def get_slot_item(
         self, slot_name: Union[DeckSlotName, StagingSlotName]
